Moved link types to model package
