+++ testit-adapter-testng/src/test/java/ru/testit/samples/SampleTest.java
@@ -6,6 +6,7 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import ru.testit.annotations.*;
 import ru.testit.models.LinkItem;
+import ru.testit.models.LinkType;
 import ru.testit.tms.client.TMSClient;
 
 public class SampleTest {
