+++ testit-java-commons/src/main/java/ru/testit/models/LinkItem.java
@@ -1,7 +1,5 @@
 package ru.testit.models;
 
-import ru.testit.annotations.LinkType;
-
 public class LinkItem
 {
     private String title;
